services/ingest/requirements.txt
# ⚠️ AUTO-GENERATED REQUIREMENTS.TXT
# This image depends on a generated requirements.txt.
# Do NOT manually edit it — instead, regenerate using:
#    ./tools/sync-requirements.sh
# Then rebuild:
#    docker compose build <service>

# SenseMy IoT Platform: Base Requirements
# Version: 1.0.1
# Last Updated: 2025-07-02

# Core Web Framework
fastapi==0.115.13

# ASGI Server
uvicorn[standard]==0.24.0

# Database Interaction
sqlalchemy==1.4.49
psycopg2-binary==2.9.10
asyncpg==0.29.0

# HTTP Client
httpx==0.28.1

# Data Validation
pydantic==2.9.2

# Logging and Monitoring
python-json-logger==2.0.7
psutil~=6.1.0

# Environment Management
python-dotenv==1.1.0

# Utilities
typing-extensions==4.14.0

# CLI Core Compatibility Fix
click==8.1.7

# Dependency Management Tools
pip-tools==7.4.1
safety==3.5.2
pip-audit==2.9.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
## Autogenerated - Do not edit this file directly
paho-mqtt==2.1.0
